1st Upload
I worked on this project during my 2009 school year. It was my first
Android application and the first one I've published on the Android
Market. I think it's a good idea I just never took the time to make it
better. Basically is a source of information to build computers on your
phone...